Prevent contbuild from causing issues
Summary:
I investigated the issue in T34928366 and found the cause is that scikit-image was updated from 0.11 to 0.13.1. This changes folder names, which can be easily fixed. However, it also breaks the patch in https://fburl.com/e0ubn164, which is harder to fix since I don't have the original file (the desired result after applying the patch).

Given that the project is no longer maintained and doesn't work in fbcode anyway (it relies on PyTorch 0.2), it doesn't seem worth fixing this issue. The only reason I am not removing the code from fbcode is because it serves as ground-truth for https://github.com/facebookresearch/adversarial_image_defenses, which we want to keep alive for a while so that people can try and reproduce the results in the corresponding paper.

This diff removes the contbuild and updates the version information in `README.md`.

Reviewed By: imisra

Differential Revision: D13222905

fbshipit-source-id: 65678d9aec8c964acdec9d49ae03bc1e84fac443